+++ src/activitypub/views.rs
@@ -7,6 +7,7 @@ use actix_web::{
     HttpResponse,
     Scope,
 };
+use serde::Deserialize;
 use serde_json::{Value as JsonValue};
 use uuid::Uuid;
 
@@ -73,16 +74,24 @@ use super::identifiers::{
 use super::receiver::{receive_activity, HandlerError};
 use super::valueflows::builders::build_proposal;
 
+#[derive(Deserialize)]
+struct ObjectQueryParams {
+    #[serde(default)]
+    fep_ef61: bool,
+}
+
 #[get("")]
 async fn actor_view(
     config: web::Data<Config>,
     db_pool: web::Data<DatabaseConnectionPool>,
     request: HttpRequest,
     username: web::Path<String>,
+    query_params: web::Query<ObjectQueryParams>,
 ) -> Result<HttpResponse, HttpError> {
     let db_client = &**get_database_client(&db_pool).await?;
     let user = get_user_by_name(db_client, &username).await?;
-    if !is_activitypub_request(request.headers()) {
+    // Do not redirect when viewing FEP-ef61 representation
+    if !is_activitypub_request(request.headers()) && !query_params.fep_ef61 {
         let page_url = get_profile_page_url(
             &config.instance_url(),
             &user.profile.username,
@@ -92,14 +101,24 @@ async fn actor_view(
             .finish();
         return Ok(response);
     };
-    let actor = build_local_actor(
-        &config.instance_url(),
-        &user,
-        false, // no FEP-ef61
-    )?;
+    let actor_value = if query_params.fep_ef61 {
+        build_local_actor_fep_ef61(
+            &config.instance_url(),
+            &user,
+            None,
+        )?
+    } else {
+        let actor = build_local_actor(
+            &config.instance_url(),
+            &user,
+            false, // no FEP-ef61
+        )?;
+        serde_json::to_value(actor)
+            .expect("actor should be serializable")
+    };
     let response = HttpResponse::Ok()
         .content_type(AP_MEDIA_TYPE)
-        .json(actor);
+        .json(actor_value);
     Ok(response)
 }
 
@@ -424,25 +443,6 @@ async fn proposal_view(
     Ok(response)
 }
 
-#[get("/fep_ef61")]
-async fn fep_ef61_actor_view(
-    config: web::Data<Config>,
-    db_pool: web::Data<DatabaseConnectionPool>,
-    username: web::Path<String>,
-) -> Result<HttpResponse, HttpError> {
-    let db_client = &**get_database_client(&db_pool).await?;
-    let user = get_user_by_name(db_client, &username).await?;
-    let actor = build_local_actor_fep_ef61(
-        &config.instance_url(),
-        &user,
-        None,
-    )?;
-    let response = HttpResponse::Ok()
-        .content_type(AP_MEDIA_TYPE)
-        .json(actor);
-    Ok(response)
-}
-
 pub fn actor_scope() -> Scope {
     web::scope("/users/{username}")
         .service(actor_view)
@@ -454,7 +454,6 @@ pub fn actor_scope() -> Scope {
         .service(subscribers_collection)
         .service(featured_collection)
         .service(proposal_view)
-        .service(fep_ef61_actor_view)
 }
 
 #[get("")]